Ran `cargo --fix`, fixed visibility of items in `built_ins/`, updated the README.

+++ src/context.rs
@@ -39,7 +39,7 @@ pub struct Context{
   // todo: Should there be a context path object?
   name   : InternedString,
   symbols: HashMap<InternedString, SymbolRecord>,
-  state: usize
+  state: u64
 }
 
 impl Context {
@@ -68,7 +68,7 @@ impl Context {
 
   // region Getters and Setters
 
-  pub fn state_version(&self) -> usize {
+  pub fn state_version(&self) -> u64 {
     self.state
   }
